Move find_section_memory_bounds back into memory_ARCH
