+++ kernel/boot_info/Cargo.toml
@@ -6,11 +6,9 @@ description = "Abstraction over multiboot2 and UEFI boot information"
 edition = "2021"
 
 [dependencies]
-log = "0.4.8"
 bitflags = "1.3"
 kernel_config = { path = "../kernel_config" }
 memory_structs = { path = "../memory_structs" }
-pte_flags = { path = "../pte_flags" }
 multiboot2 = { version = "0.14", optional = true }
 
 [dependencies.uefi-bootloader-api]
