+++ kernel/memory_x86_64/Cargo.toml
@@ -6,15 +6,12 @@ version = "0.1.0"
 
 [dependencies]
 x86_64 = "0.14.8"
+log = "0.4.8"
 
-[dependencies.boot_info]
-path = "../boot_info"
-
-[dependencies.pte_flags]
-path = "../pte_flags"
-
-[dependencies.memory_structs]
-path = "../memory_structs"
+boot_info = { path = "../boot_info" }
+pte_flags = { path = "../pte_flags" }
+kernel_config = { path = "../kernel_config" }
+memory_structs = { path = "../memory_structs" }
 
 [lib]
 crate-type = ["rlib"]
